+++ plugin.php
@@ -1,7 +1,7 @@
 <?php
 /**
- * Plugin Name: TMS Materials
- * Plugin URI: https://github.com/devgeniem/tms-plugin-materials
+ * MaterialsPlugin Name: TMS Materials
+ * MaterialsPlugin URI: https://github.com/devgeniem/tms-plugin-materials
  * Description: TMS Materials
  * Version: 0.0.1
  * Requires PHP: 7.4
@@ -13,7 +13,7 @@
  * Domain Path: /languages
  */
 
-use TMS\Plugin\Materials\Plugin;
+use TMS\Plugin\Materials\MaterialsPlugin;
 
 // Check if Composer has been initialized in this directory.
 // Otherwise we just use global composer autoloading.
@@ -28,15 +28,15 @@ $plugin_version = $plugin_data['Version'];
 $plugin_path = __DIR__;
 
 // Initialize the plugin.
-Plugin::init( $plugin_version, $plugin_path );
+MaterialsPlugin::init( $plugin_version, $plugin_path );
 
 if ( ! function_exists( 'tms_plugin_materials' ) ) {
     /**
      * Get the plugin instance.
      *
-     * @return Plugin
+     * @return MaterialsPlugin
      */
-    function tms_plugin_materials() : Plugin {
-        return Plugin::plugin();
+    function tms_plugin_materials() : MaterialsPlugin {
+        return MaterialsPlugin::plugin();
     }
 }